feat(library): refactor
