+++ mark_authenticator/src/main/java/com/fadlurahmanfdev/mark_authenticator/api/callback/MarkAuthenticationCallback.kt
@@ -1,6 +1,6 @@
 package com.fadlurahmanfdev.mark_authenticator.api.callback
 
-import com.fadlurahmanfdev.mark_authenticator.core.exception.MarkAuthenticatorException
+import com.fadlurahmanfdev.mark_authenticator.domain.exception.MarkAuthenticatorException
 import javax.crypto.Cipher
 
 /**
