+++ mark_authenticator/src/main/java/com/fadlurahmanfdev/mark_authenticator/domain/enums/MarkAuthenticationType.kt
@@ -1,4 +1,4 @@
-package com.fadlurahmanfdev.mark_authenticator.core.enums
+package com.fadlurahmanfdev.mark_authenticator.domain.enums
 
 internal enum class MarkAuthenticationType {
     BIOMETRIC_WEAK,
