+++ mark_authenticator/src/main/java/com/fadlurahmanfdev/mark_authenticator/domain/exception/MarkAuthenticatorException.kt
@@ -1,4 +1,4 @@
-package com.fadlurahmanfdev.mark_authenticator.core.exception
+package com.fadlurahmanfdev.mark_authenticator.domain.exception
 
 /**
  * Domain exception exposed to consumer applications.
